chore(sessions): break out sessions manager code
